Extracted zooming into a separate wrapper component
