+++ client/app/scripts/charts/nodes-chart-elements.js
@@ -13,10 +13,10 @@ import {
 
 class NodesChartElements extends React.Component {
   render() {
-    const { layoutNodes, layoutEdges, selectedScale, transform, isAnimated } = this.props;
+    const { layoutNodes, layoutEdges, selectedScale, isAnimated } = this.props;
 
     return (
-      <g className="nodes-chart-elements" transform={transform}>
+      <g className="nodes-chart-elements">
         <NodesChartEdges
           layoutEdges={layoutEdges}
           selectedScale={selectedScale}
